+++ ssg-md/src/test/scala/ssg/md/ext/gfm/strikethrough/test/ComboSubscriptSpecTest.scala
@@ -28,7 +28,7 @@ final class ComboSubscriptSpecTest extends RendererSpecTestSuite {
 object ComboSubscriptSpecTest {
   val SPEC_RESOURCE:     String           = "/ssg/md/ext/gfm/strikethrough/test/ext_subscript_ast_spec.md"
   val RESOURCE_LOCATION: ResourceLocation = ResourceLocation.of(classOf[ComboSubscriptSpecTest], SPEC_RESOURCE)
-  val OPTIONS:           DataHolder       = new MutableDataSet().set(HtmlRenderer.INDENT_SIZE, 0).set(Parser.EXTENSIONS, Collections.singleton(SubscriptExtension.create())).toImmutable()
+  val OPTIONS: DataHolder = new MutableDataSet().set(HtmlRenderer.INDENT_SIZE, java.lang.Integer.valueOf(0)).set(Parser.EXTENSIONS, Collections.singleton(SubscriptExtension.create())).toImmutable()
 
   val OPTIONS_MAP: java.util.Map[String, DataHolder] = {
     val map = new HashMap[String, DataHolder]()
